pyproject.toml
[build-system]
requires = [
    "Nuitka",
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
name = "TiLiA"
version = "0.5.15.1"
description = "A GUI for creating and visualizing annotations over audio and video files."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "colorama~=0.4.6",
    "httpx~=0.28.1",
    "isodate~=0.7.2",
    "lxml>=5.3.0",
    "music21>=9.1",
    "numpy~=1.26.4",
    "platformdirs>=4.0.0",
    "prettytable>=3.7.0",
    "python-dotenv>=1.0.1",
    "pypdf>=6.0.0",
    "PySide6>=6.8.0",
    "PySide6_Addons>=6.8.0",
    "PySide6_Essentials>=6.8.0",
    "sentry-sdk>=2.21.0",
    "soundfile~=0.13.1",
    "TiLiA",
    "tomli; python_version < '3.11'"
]

[[project.authors]]
name = "Felipe Defensor et al."
email = "tilia@tilia-app.com"

[project.gui-scripts]
tilia-desktop = "tilia.__main__:boot"

[project.urls]
Homepage = "https://tilia-app.com"
Repository = "https://github.com/TimeLineAnnotator/desktop"

[dependency-groups]
build = [
    "build",
    "ImageIO; sys_platform == 'darwin'",
    "Nuitka",
    "pyyaml",
    "wheel"
]
ci-tests = [
    "coverage>=7.0.0",
    "coverage-badge>=1.0.0",
    "flake8>=7.0.0",
    "pytest-cov>=6.0.0",
]
dev = [
    "icecream>=2.1.0",
    "pre-commit",
]
testing = [
    "pytest>=8.0.0",
    "pytest-env>=1.1.5",
]

[tool.coverage.run]
source = ["tilia"]
omit = ["tilia/dev/*"]

[tool.nuitka]
enable-plugins = ["pyside6"]
user-package-configuration-file = "tilia.nuitka-package.config.yml"
noinclude-qt-translations = true
include-qt-plugins = ["multimedia"]
nofollow-import-to = ["*.tests", "*.test"]
assume-yes-for-downloads = true
# deployment = true

[tool.pytest.ini_options]
env = [
    "ENVIRONMENT=test",
    "QT_QPA_PLATFORM=offscreen",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = [
    "build", "build.*",
    "dist",
    "docs", "docs.*",
    "paper", "paper.*",
    "scripts",
    "tests", "tests.*",
    "venv.*", "venv"
]

[tool.setuptools.package-data]
tilia = [
    "../.tilia.env",
    "../tilia.nuitka-package.config.yml", # because we build from sdist
    "../LICENSE",
    "ui/img/*",
    "ui/fonts/*",
    "media/player/youtube.html",
    "media/player/youtube.css",
    "parsers/score/svg_maker.html",
    "parsers/score/timewise_to_partwise.xsl",
]
